backend/requirements.txt
torch
torchvision
opencv-python
Pillow
requests
numpy
fastapi
uvicorn
